Refactor commands pertaining to xml manifest
